chore (dashboard): refactore dashboard page
